requirements.txt
python-dotenv>=0.11.0
tweepy>=3.8.0
xmltodict>=0.12.0
pandas>=1.0.1
Pillow==9.4.0
imgkit==1.2.2
openai
beautifulsoup4
lxml
